+++ gbus/tx/timeout.go
@@ -1,4 +1,4 @@
-package saga
+package tx
 
 import (
 	"database/sql"
